Separate actions, reducers and store
