+++ redux/actions.js
@@ -1,0 +1,14 @@
+// Action Types
+const NAVIGATE = 'NAVIGATE'
+
+// Action creators
+const navigateTo = (step) => {
+  return {
+    type: NAVIGATE,
+    step: step
+  }
+}
+
+export {
+  navigateTo
+}
